Add alternative layout without content in bloks
This alternative layout was implemented in case the content is non-essential to the consumption of the content on the page and we do not want it to be visible for search engines.
Also with this commit message, I am trying to improve my skill of writing an understandable and readable commit message.